[Gunveer | Harshit] Add query-executor api and its test
