+++ server/app/api/v1/query.py
@@ -63,48 +63,6 @@ async def query(
         status_code=status.HTTP_200_OK,
     )
 
-@router.post("/query-executor")
-async def query_executor(
-    db_id: Annotated[str, Form()],
-    sql_query: Annotated[str, Form()],
-    db: Session = Depends(get_db),
-    current_user: dict = Depends(get_current_user),
-) -> JSONResponse:
-    try:
-        database_connection_string = crud_user_database.get_by_id(db, db_id).connection_string
-        if database_connection_string:
-            engine = create_engine(database_connection_string)
-            with engine.connect() as connection:
-                result = connection.execute(text(sql_query))
-                result = result.mappings().all()
-                result_in_json_format = jsonable_encoder(result)
-                result_in_2d_array = query_service.convert_to_2d_array(result_in_json_format)
-                logger.info(
-                    "Query {} executed successfully for user {} and database {}".format(
-                        sql_query, current_user.id, db_id
-                    )
-                )
-                
-
-    except Exception as e:
-        logger.error(
-            "Error while executing query {} for user {} and database {}. Error is {}".format(
-                sql_query, current_user.id, db_id, e
-            )
-        )
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error",
-        )
-
-    return JSONResponse(
-                    content={
-                        "message": "Query executed successfully",
-                        "data": {"sql_query": sql_query, "query_result": result_in_2d_array},
-                    },
-                    status_code=status.HTTP_200_OK,
-    )
-
 @router.get("/queries")
 async def get_queries(
     db_id: str,
