+++ server/app/tests/test_api/v1/test_query.py
@@ -1,4 +1,3 @@
-import uuid
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 from unittest.mock import patch
@@ -6,8 +5,6 @@ from app.models.user import User as UserModel
 from app.crud.crud_user_database import crud_user_database
 from app.crud.crud_query import crud_query
 from app.schemas.user_database import UserDatabase as UserDatabaseSchema
-from app.clients.openai_client import OpenAIClient 
-from app.crud.crud_user_database import CRUDUserDatabase
 from app.schemas.query import Query as QuerySchema
 
 mock_openai_embedding_response = {
@@ -169,45 +166,3 @@ def test_query(
             ],
         },
     }
-        
-        
-   
-
-def test_query_by_id(
-    client: TestClient, db: Session, valid_jwt: str, valid_user_model: UserModel
-) -> None:
-    database_obj = crud_user_database.create(
-        db=db,
-        user_database_obj=UserDatabaseSchema(
-            name="Test",
-            user_id=valid_user_model.id,
-        ),
-    )
-    query_obj = crud_query.create(
-        db=db,
-        query_obj=QuerySchema(
-            nl_query="show me the table schema of the table employee",
-            user_database_id=database_obj.id,
-            sql_query="SELECT * FROM employee",
-        ),
-    )
-    response = client.get(
-        f"api/v1/queries/{query_obj.id}",
-        headers={"Authorization": f"Bearer {valid_jwt}"},
-    )
-    assert response.status_code == 200
-    assert response.json() == {
-        "message": "Query fetched successfully",
-        "data":{
-            "query": {
-                "id": str(query_obj.id),
-                "nl_query": "show me the table schema of the table employee",
-                "sql_query": "SELECT * FROM employee",
-                "user_database_id": str(database_obj.id),
-                "created_at": query_obj.created_at.isoformat(),
-                "updated_at": query_obj.updated_at.isoformat(),
-        }
-            
-        }
-    }
-
